[#318] Create scandir mod
